+++ python/cudf/cudf/tests/test_sorting.py
@@ -271,14 +271,14 @@ def test_dataframe_scatter_by_map(map_size, nelem):
     df["a"] = np.random.choice(strlist[:map_size], nelem)
     df["b"] = np.random.uniform(low=0, high=map_size, size=nelem)
     df["c"] = np.random.randint(map_size, size=nelem)
-    df["d"] = df["a"]._column.as_categorical_column(np.int64)
+    df["d"] = df["a"]._column.as_categorical_column(np.int32)
 
     def _check_scatter_by_map(dfs, col):
         assert len(dfs) == map_size
         nrows = 0
         for df in dfs:
             nrows += len(df)
-            assert df[col].astype(np.int64).nunique() <= 1
+            assert df[col].astype(np.int32).nunique() <= 1
         assert nrows == nelem
 
     _check_scatter_by_map(df.scatter_by_map("a", map_size), "a")
